Hipcheck version 3.1.0-alpha.0
This commit encompasses work that's happened over a span of 3 years,
from May 9th, 2019 onward. Hipcheck started with a small team working
on the problem of trying to assess software dependencies for supply
chain risk. Team size and funding fluctuated regularly, with the team
periodically having only a single developer. Eventually, we were able
to grow it into something bigger, and we're proud of the work we've
done to make a tool that helps projects understand risks. We're also
very happy to see the way the software supply chain security field
has grown and developed since we first started working on this
problem.

We're releasing Hipcheck with the intent to take the best parts of
it and integrate them into Scorecard, an effort from the Open Source
Security Foundation with the same goals. We don't intend to further
develop Hipcheck, and don't recommend using it.

All of that said, enormous thanks to every contributor to Hipcheck
during its lifespan.

Co-authored-by: Andrew Lilley Brinker <abrinker@mitre.org>
Co-authored-by: Michael Chernicoff <mchernicoff@mitre.org>
Co-authored-by: Rebecca Dornin <rdornin@mitre.org>
Co-authored-by: Ian Kretz <ikretz@mitre.org>
Co-authored-by: Cal Stepanian <cstepanian@mitre.org>
Co-authored-by: Gananand Kini <gkini@mitre.org>
Co-authored-by: Victor Sun <vsun@mitre.org>
Co-authored-by: James Frank Murphy <jfmurphy@mitre.org>
Co-authored-by: Steve Battista <sbattista@mitre.org>
Co-authored-by: Cathleen Zhang <czhang@mitre.org>
Co-authored-by: Deyuan Li <deyuanli@mitre.org>