chore: Upgrade version-compatible deps to latest.
Some, but not all, of our out-of-date dependencies are only a minor
or pre-1.0 patch release away from their latest version, which means we
can upgrade to them with a simple `cargo update`. This commit is the
result of doing that update, and the remaining out-of-date dependencies
are one which will likely require more work as the new versions should
not be expected to be compatible with the versions we're using today.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>