fix: Fix missing actions on Clap config
Clap defines the CLI of Hipcheck, and in the transition to the latest
version of Clap I missed that some arg actions are no longer inferred
and need to be set explicitly. This commit adds those missing actions.

Other options do appear to arise though, so the transition to Clap 4 is
not yet done.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>