Upgrade Hipcheck dependencies (#21)
* chore: Upgrade version-compatible deps to latest.

Some, but not all, of our out-of-date dependencies are only a minor
or pre-1.0 patch release away from their latest version, which means we
can upgrade to them with a simple `cargo update`. This commit is the
result of doing that update, and the remaining out-of-date dependencies
are one which will likely require more work as the new versions should
not be expected to be compatible with the versions we're using today.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Upgrade "env_logger" to latest version.

This commit upgrades env_logger to the latest version, and also resolves
some compilation errors that arose alongside that upgrade.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Update 'toml' to latest in xtask.

This commit bumps the 'toml' crate to the latest version.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Upgrade 'which' to latest.

This commit bumps the 'which' crate dependency in 'xtask' to the latest
version.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Upgrade 'open' crate in 'xtask' to latest

The commit upgrades the `open` crate dependency in the `xtask` crate to
the latest version.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Upgrade 'clap' in 'xtask' to latest.

This commit upgrades the `clap` crate dependency in the `xtask` crate to
the latest version.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Upgrade 'which' in 'hc_common' to latest.

This commit bumps `which` in `hc_common` to the latest version.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Update `ordered-float` in `hc_common`.

This commit updates `ordered-float` in `hc_common` to the latest
version.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Updates 'dirs' in 'hc_modules'

This commit updates the 'dirs' crate dependency in 'hc_modules' to the
latest version.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Update 'dirs' in 'hc_session'

This commit updates the 'dirs' dependency in 'hc_session' to the latest
version.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Upgrade 'dirs' to latest in 'hc_npm'

This commit bumps the 'dirs' crate dependency to the latest version in
the 'hc_npm' crate.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Update 'dirs' in 'hc_git' to latest.

This commit bumps the 'dirs' crate dependency in 'hc_git' to the latest
version.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Bump 'toml' in 'hc_filesystem' to latest.

This commit bumps the 'toml' crate dependency in 'hc_filesystem' to the
latest version.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Bump 'smart-default' in 'hc_config' to latest

This commit bumps the 'smart-default' crate dependency in 'hc_config' to
the latest version.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Bump 'toml' in 'hc_pm' to latest

This commit bumps the 'toml' crate in `hc_pm` to the latest version.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Remove unused deps in 'hc_pm'

This commit removes 'clap', 'reqwest', and 'http' as dependencies of the
`hc_pm` crate. They were holdovers from a previous version of the code
which ran as its own binary, and aren't used in the code today.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Upgrade nom to 7.x in hc_git.

This commit upgrades the `nom` crate from 6.x to 7.x as a dependency of
`hc_git`.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Update 'clap' to latest.

This commit updates 'clap' to the latest version as a dependency of the
'hipcheck' crate.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Fixup cargo clippy lints.

This commit fixes a bunch of open cargo clippy lints.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Move from tempdir to tempfile in hc_session

Turns out the 'tempdir' crate has been deprecated in favor of
'tempfile'. This commit moves to the new crate.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Update patch deps

This commit runs 'cargo update' to update a bunch of patch dependencies
to the latest in the lockfile.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Bump graphql_client to latest.

This commit bumps the `graphql_client` library dependency in `hc_github`
to the latest version.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* chore: Upgrade 'ureq' to 2.x

This commit upgrades the `ureq` crate to version 2.x, and updates our
code to ensure authentication is still done on all requests.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* fix: Fix missing actions on Clap config

Clap defines the CLI of Hipcheck, and in the transition to the latest
version of Clap I missed that some arg actions are no longer inferred
and need to be set explicitly. This commit adds those missing actions.

Other options do appear to arise though, so the transition to Clap 4 is
not yet done.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

* fix: Replace broken unwrap in fuzz HTTP request.

This commit fixes the logic for processing responses from the GitHub
CodeSearch API which power the fuzzing analysis. Previously, it was
panicking if the request failed, and now it will error gracefully and
not panic the program.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>

---------

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>