feat: Define devcontainer config
This commit defines a devcontainer configuration for Hipcheck, to enable
contributors to Hipcheck to spin up a devcontainer that has what they
need to work on the project.

Right now that means it installs the latest Rust toolchain, the latest
Long-Term Support (LTS) Node.js toolchain, and then sets up OpenSSL dev
headers and the NPM module-deps package.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>