fix: Remove broken Windows import.
This is why we have cross-platform CI testing haha.

This commit removes a broken import in some Windows-specific code. We'd
previously done a weird thing of re-exporting external dependencies
under crate-local pathnames. We no longer do that, but had an import
laying around still that expected this kind of structure.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>