chore: Removed dead code.
This commit removes all "#[allow(unused)]` annotations and all the dead
code that removal of those annotations revealed. This is code that isn't
used today and which we don't expect to use in the future, though of
course it exists in the project history so we can get it back if we
really need to.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>