fix: Fix double-version command in `xtask`
This commit fixes a runtime error that was stopping `cargo xtask` from
running at all. Basically, we define a custom `--version` command, which
was conflicting with the one automatically generated by `clap`,
resulting in a runtime failure.

This command just adds a line telling `clap` not to generate its own
`version` flag, resolving the error and allowing `xtask` to run as
expected.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>