chore: Added basic community docs.
This commit adds some basic "community" docs like a code of conduct
and contributing file.

It also fixes the file extension for the license (it's not really
markdown, so `.md` wasn't right), and adds a little explanation of the
purpose of the helper script at the top just for clarity's sake.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>