feat(xtask): add `xtask changelog` command.
This commit introduces a new `xtask changelog` command to `cargo xtask`,
to help produce automated changelogs which can be the basis for updating
`CHANGELOG.md`. This new command _does not_ overwrite `CHANGELOG.md`,
but instead writes its auto-generated changelog to `CHANGELOG-NEXT.md`,
which is also marked in `.gitignore` so it doesn't get checked in. The
goal here is that this automation will _help_ write changelogs for each
new version, but doesn't turn the changelog into a fully-automated
system, as we still want to be able to have human-written release notes.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>