chore: Clean up CLI implementation.
This commit is a pretty substantial internal refactoring of the CLI
implementation, but is purposefully designed not to break any backwards
compatibility. A number of existing flags have been deprecated and
removed from the CLI help text, but in fact are still present and work
as they did before.

This commit also introduces `hipcheck-macros`, a procedural macro helper
crate for Hipcheck.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>