feat: Add a new Hipcheck website
This is a first-pass implementation of a new Hipcheck website. It is
as-of-yet incomplete, but will host installation instructions, official
documentation, Requests for Discussion, and more.

This does _not_ include the logic for publishing the site to GitHub
Pages using a GitHub Action, which will be handled in a separate commit.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>