fix: Provide Zola version directly.
The prior version of this tried to use an environment variable to set
the version of Zola to install. This failed in CI (#1). I _think_ it
failed because "use/with" directives are at the GitHub Actions level,
where environment variables are not expanded, rather than the
embedded-shell level of "run" directives where environment variables
_are_ expanded, but I don't feel confident enough in my understanding of
GitHub Actions to be sure.

At any rate, now I'm not trying to be clever, and just set the version
number directly.

1: https://github.com/mitre/hipcheck/actions/runs/9726634299/job/26845303002

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>