feat: Added `--env` flag to `xtask site serve`
This adds a new `--env` flag to the `xtask site serve` subcommand that
lets you specify if you're wanting to run in production or development
mode. The only thing this changes at the moment is the `base_url` of the
site, intended so you can locally test the production configuration of
the deployment. By default, it'll use `--env dev`, which sets the
`base_url` to `localhost`.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>