feat: Completed MVP of Hipcheck Complete Guide
This is a first-pass completion of the Hipcheck Complete Guide
documentation, intended to provide a thorough reference to the concepts
and functionality of Hipcheck. I'm sure there will be many more
improvements to make in the future, but with this I do now feel
comfortable moving forward with this documentation and further
publication of the site (pending some design improvements to go along
with these content improvements).

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>