fix: Fix broken website links in production.
Previously, by hardcoding a number of links, we were breaking in
production because the production site is served under a sub-path of the
`https://mitre.github.io` host: `hipcheck/`. This commit fixes many of
these broken links, including those on the homepage and in the
navigation and footer, as I now better understand Zola's mechanism for
linking to pages within the site in a way that always ensures proper
handling of the site's `base_url`.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>