feat: Added RFD #4
This RFD describes in detail the planned design of the Hipcheck plugin
system. This is still a draft design, but marks the first reasonably
complete description of the planned design of this system.

The goal here is to solicit feedback from the Hipcheck team and outside
interested parties to ensure we can make this thing usable for plugin
developers, and that we've identified roadblocks and design problems in
advance of implementation as much as possible.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>