chore: Integrated policy expressions code
This is still a work-in-progress, and nothing is actually used
yet, but this at least moves over the draft implementation of policy
expressions from the external experimental repo I'd made to enable
integration with Hipcheck.

The other thing I did was organize the Cargo.toml file for the hipcheck
crate, particularly to put all dependencies in alphabetical order, as
they had been at one point previously.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>