fix: Fix broken containerfile.
The containerfile used for 3.4.0 and 3.5.0 was quietly broken and
deploys to Docker Hub were failing. This commit _should_ fix the
brokenness based on local testing. The file has also been shortened,
removing comments which were not needed or no longer accurate. Commands
in the `RUN` directive are also now joined by `&&` instead of `;` to
ensure early termination if an error arises.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>