feat: Update proto def to pass 'buf lint'
This commit modifies the protobuf definition for Hipcheck plugins to meet the
expectations of 'buf lint', which checks protobuf definitions against a set
of best practices. The main goal here is to make the service definition more
defensive against forward-compatibility issues.

This also includes updates to the code that interacts with the generated code
to make sure it still compiles.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>