chore: Remove unused version handling
Previously (before open sourcing Hipcheck) we'd update the `Cargo.toml`
file for the Hipcheck crate after releasing a new version to reflect that
we're now working on a pre-release of the _next_ version, and we had special
handling in Hipcheck itself to identify at build time when we were building
one of these pre-release versions and add the commit hash of the repository
itself to Hipcheck's version. This was cool, and the idea was to make it
obvious when you were running a pre-release build of Hipcheck and _which_
pre-release build you were running.

However, when we transitioned to open source, we changed how releases
happen, and stopped doing the pre-release marking, which means this special
handling code hadn't been used at all since then. What's more, it's an extra
piece of complexity that we've been carrying around. Rather than try to
once again change our release flow, I'm removing the special handling code
since it's unused.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>