chore: Create new 'init' top-level module
Over time we've accrued a number of pieces of logic for pieces of global
state that needs to be initialized right at the start of execution. These
pieces all init-related but scattered around modules in a way that meant
more clutter at the top-level of the crate, and a general lack of clarity
about how init was done. This refactoring groups all init-related logic
under an `init` function in the `init` module, which is just called right
at the top of the `main` function.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>