fix: Broken release build for Hipcheck.
Turns out Cargo doesn't like when materials required to build a crate are
outside of the crate manifest directory. In this case, I'm talking about
the protobuf definition used for the Hipcheck plugin gRPC service.

There's a Cargo issue open about it, but the gist of it is that they
don't really want people doing what we were trying to do, having materials
in an outer directory.

There may be better solutions, but moving the files is the fastest one
for now.

Here's the issue: https://github.com/rust-lang/cargo/issues/3946

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>