fix: Install the protobuf compiler in release CI
Turns out we'd forgotten to update the cargo-dist configuration to install
the protobuf compiler for our release builds in CI. This commit _should_
fix it, though I'm not able to test it until it's deployed and I try to cut
a new release.

The basic idea here is that cargo-dist provides configuration items to
use platform-specific package managers to install needed packages. In our
case, funnily enough, the protobuf compiler package has a different name
on every platform.

For the moment I am not tying us to a specific version, just using the `*`
to indicate we should always be using latest. If that becomes an issue in
the future we can tighten it up.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>