feat: Add `Analysis::Plugin` variant.
This defines a new variant of the `Analysis` type used in reporting,
called `Plugin`, which encodes the data necessary to print the results
of plugins. In the future this will be the _only_ variant of the enum,
and then we can get rid of the enum entirely, but for now this is a
stop-gap to bridge between the old non-plugin execution and the new
plugin system.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>