chore: Revert 'cargo-dist' workaround
We hit a bug in 'cargo-dist' yesterday that it turns out was already fixed in
the latest version, but before we figured that out we implemented a
workaround to enable us to get 3.6.2 out the door. Now that we know the
latest 'cargo-dist' version doesn't need the workaround, we can revert it
and use 0.22.1.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>