feat: Evaluate JSON Pointers in Policy Expressions
This commit replaces the previous preprocessor-based implementation of
JSON Pointers with one that natively runs in the Policy Expression
`Executor`.

It replaces the `context` parameter on `run()` and `parse_and_eval()`
with a new field on the `Env` struct. Now, calling code must pass the
JSON context while creating the `Env`.

It also fixes a bug that the previous version had but which was hidden
by the intermediate step of serializing to text: numbers in the JSON
that could be represented as integers were being forced to f64. This is
a problem for Policy Expressions, which distinguishes between Ints and
Floats and does not currently have any casting or conversion between
them. Since the previous version went through Rust's native printing of
f64, and that printing code will display a float without a decimal point
if it is exactly equal to an integer, the bug did not manifest.

This version goes directly from the Expr returned by
`json_to_policy_expr()` to the Executor, so it includes a fix to return
Primitive::Int values where appropriate.