feat: Simplify project `Containerfile`
Previously, our `Containerfile` would try to build the project as it exists
in the currently checked-out local repo, and package that up. This provided
some complexity for publication of new versions, as publishing container
images is the _last_ step of our release process, and we _have_ seen instances
where all prior steps succeeded and a release is now "out" but the container
build failed, and so publication to Docker Hub failed as well.

This commit changes the logic of the `Containerfile` to instead just install
a prebuilt binary from our GitHub Releases install script. This means we don't
worry about building Hipcheck inside of the container.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>