fix: Fix broken `Containerfile`
It turns out the prior version of the `Containerfile` was broken,
as the install script for Hipcheck wouldn't actually work due to a
missing `xz-utils` package. This was obscured by my own networking
issues, causing me to inadequately test the prior change.

One other change this makes is that it splits the `RUN` commands
up with a `USER` command, so the first `RUN` is done as root to set
up the system, and the second `RUN` is done as `hc_user` to install
Hipcheck.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>