feat: add --arch flag to `hc check` subcommand to allow architecture detection override
We had previously encountered problems with cross-compilation to
niche-but-valid architectures, where the compile-time target detection
would fail, causing `CURRENT_ARCH` to be `()` instead of a valid arch
enum variant. We changed this field to be type `Option<SupportedArch>`
and also support a user-provided `--arch` flag to override the detected
arch at runtime if the user has good reason to do so.