feat: Reorg libs and update protobuf defs
This commit moves the two Hipcheck "macro" libraries (`hipcheck-macros` and
`hipcheck-sdk-macros`) under the `libraries/` folder, which going forward will
be used as the directory for any support library crates we define in the
Hipcheck project.

It also deduplicates the protobuf definition files, ensuring that all users of
them are synced with the central definition, and updates them to comply to
`buf`'s recommended best practices for protobuf files. This includes updates
to all users of the definitions to comply with the new specification. Note that
these changes are backwards incompatible, as they involve changing the names
of existing calls in the gRPC interface.

This is a breaking change, but given that the plugin interface is still
experimental and without real operational users, this is the right time to
get the interface definition right without letting ourselves be locked in to
compatibility.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>