feat: Evaluate JSON Pointers using late binding
This commit replaces the previous preprocessor-based implementation of
JSON Pointers with one that natively runs in the Policy Expression
`Executor`.

It uses the ExprMutator visitor trait to evaluate JSON Pointers in two
steps: first, pointers are looked up but the original pointer string is
preserved. Then, the looked-up value is inserted.  This structure is
meant to allow for better error messages and user-facing explanations of
why a Policy Expression was evaluated a given way.