feat: rust sdk sends error gRPC message on internal error
This commit addresses an outstanding bug in the plugin Rust SDK.
Previously, the SDK would print when an erorr occurred during
`handle_session()`, which executes a query, but it did not send an
"error" gRPC message to Hipcheck core, causing analysis to hang while
waiting for a message that would not come. This commit cleans up the
handling and logging of plugin-side internal errors.