fix: Make parser handle meta lines in chunk
The previous parser for Git chunks would fail to properly break
up chunks with any "meta lines" (my term for them) starting with
a `\` instead of `+` or `-`. This update to the parser means the
parser will now correctly handle those lines, but won't include
them in the patch string it constructs, as they're meta-commentary
on the diff itself and not part of the edits.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>
Co-authored-by: Julian Lanson <jlanson@mitre.org>