fix: `hc cache` deletion logic
The `cache delete` subcommand had a bug in how it was sorting entries to
be deleted or preserved, so that `-s all` would actually choose to keep
all instead of delete all entries. Also updated the per-entry deletion
logic to also clean up the parent dir (corresponding to github owner)
if it was made empty by deleting the child dir.