feat: Improve reporting messages.
This commit has two benefits, one is to eliminate a lot of the old
reporting code which special-cased the legacy analyses, now that those
have been entirely excised. The other is to improve the error reporting
in both the success and failure cases.

There is still more that could be done to make these messages more
explanatory, but this is a start.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>