fix: performance concerns in the git plugin's contributors_for_commit
Although the git plugin's `RawCommit` struct contains author and
committer information, the `Commit` struct does not. For looking up the
contributors to a commit we were doing a lot of expensive and wasteful
computation to re-engineer what is already available in `RawCommit`.
This commit exposes more functionality to the query endpoint functions
to get `RawCommit` collections so we spend less time looking for
information that was already there before we discarded it.