fix: graceful closing of other plugins on initialization failure
Now that we parse all default expressions from plugins, we discovered
that the `activity` plugin had a malformed expression, and when the
parsing failed and `hc` exited, other plugin processes were left
orphaned. This commit makes sure we use up the JoinSet that performs
plugin initialization, thus trying to initialize all plugins instead of
stopping as soon as the first error occurs. By doing so, we prevent
orphaned processes and report all errors which makes for a better user
experience.

Signed-off-by: jlanson <jlanson@mitre.org>