chore: Eliminate CHANGELOG.md
We've so far maintained a CHANGELOG.md file in the root of the
repository, but over time the number of distinct releasable things that
we manage has increased. Each of these is separately versioned.

Going forward, we'll be publishing release announcements on the Hipcheck
blog, and for all prior information the data from the CHANGELOG was
being pulled into GitHub Releases anyway, so we won't lose it by
removing it from the repo.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>