feat: add `manifest` xtask to parse GitHub Release API output
This commit is an initial part of a larger effort to develop an `xtask`
command to automate plugin download manifest file updates when new
plugin releases are made. This commit simply calls the GitHub release
API for hipcheck, and parses and manipulates the returned JSON into
DownloadManifestEntry objects that can be compared against the entries
in existing download manifest files.

Signed-off-by: jlanson <jlanson@mitre.org>