fix: get proper version from download manifest and spawn with canonical path
Discovered a bug in the entry selection process from the plugin's download
manifest, such that the desired version was not checked and the first
entry with a matching arch was returned. Additionally, downloaded
plugin.kdl files have entrypoints that are relative to the parent dir,
but Hipcheck core does not change working directory for the process so
we don't find the right plugin binary. This commit prepends the proper
plugin cache dir to PATH when resolving the binary to execute so that
the `git` plugin's `git` binary will be found before the actual version
control software. This commit also passes entrypoint args properly to
the plugin's Command object.

Signed-off-by: jlanson <jlanson@mitre.org>