feat: Default to policy file at standard location
Resolves #582
Resolves #682

Hipcheck now defaults to using a policy file called Hipcheck.kdl in the
default config dir.
It will fall back to loading the legacy TOML format if that fails.

If the user explicitly calls for --policy or --config, Hipcheck will
only attempt that source of configuration, not falling back to the
other.

Add a new Provenance enum for tracking whether a setting was set
explicitly by a user or is from defaults.

Add a new ConfigMode enum to make fewer invalid states representable.

Signed-off-by: Cal Stepanian <61707847+cstepanian@users.noreply.github.com>