feat: Enable merge queues for CI actions.
Make all current PR-enabled CI actions also run within merge groups.
This is part of introducing merge queues to the Hipcheck CI process,
which should help ensure we avoid breakage where changes pass on
a branch in CI, but then fail post-merge because of how the change
interacts with changes on `main` since the feature branch for the
PR diverged.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>