chore: CI workflows always run, jobs do filtering
Merge queues trigger jobs based on the "required jobs" section of a
branch's merge protections. Unfortunately, merge queues cannot detect
when a workflow is "skipped", and so expects all of the required jobs to
return success. Our current workflows decide at the top-level whether to
run based on the modified files in the PR, so sometimes these jobs
considered "required" by the merge queue do not fire, causing a
deadlock. This PR moves the filtering into a job of the workflow, so now
the workflow always "runs", but the relevant job will not unless any
files in the given paths are modified.

Signed-off-by: jlanson <jlanson@mitre.org>