chore: Try reducing Windows Actions times.
Windows actions on GitHub are slow, often twice as slow as the
same actions on Ubuntu or macOS. There's some suggestions
online that this can be due to slow disk performance when writing
to the C drive, so this commit moves some key tools to the
D drive instead. We'll have to test them in CI to see if this
actually improves performance.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>