fix: broken 'rust-toolchain' GitHub Action
I'd amended all our GitHub Actions dependencies to use pinned
hashes for specific commits, but it turns out that the
"dtolnay/rust-toolchain" Action uses the target hash as the
way to know what toolchain you want (I wish it wouldn't, and
would instead use an argument like a normal GitHub Action, but
alas). This reverts the change for stable so it will work
again.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>